feat: Long 값인 runner_id를 가지고 Authentication을 만들 수 있도록 구현

+++ src/main/java/com/aetheri/application/service/converter/RunnerAuthenticationConverter.java
@@ -18,7 +18,7 @@ public class RunnerAuthenticationConverter {
         );
     }
 
-    public static Authentication toAuthentication(Integer runnerId) {
+    public static Authentication toAuthentication(Long runnerId) {
         return new UsernamePasswordAuthenticationToken(
                 runnerId,               // principal
                 null,                   // credentials
